Added delete buttons, removed delete cards
