+++ src/components/[guild]/RolesByPlatform/components/RoleListItem/components/EditRole.tsx
@@ -1,13 +1,12 @@
 import {
   Button,
-  Divider,
   Drawer,
   DrawerBody,
-  DrawerCloseButton,
   DrawerContent,
   DrawerFooter,
   DrawerOverlay,
   Heading,
+  HStack,
   Icon,
   IconButton,
   useBreakpointValue,
@@ -21,7 +20,7 @@ import Description from "components/create-guild/Description"
 import LogicPicker from "components/create-guild/LogicPicker"
 import NameAndIcon from "components/create-guild/NameAndIcon"
 import Requirements from "components/create-guild/Requirements"
-import DeleteRoleCard from "components/[guild]/edit/[role]/DeleteRoleCard"
+import DeleteRoleButton from "components/[guild]/edit/[role]/DeleteRoleButton"
 import useEditRole from "components/[guild]/edit/[role]/hooks/useEditRole"
 import usePersonalSign from "hooks/usePersonalSign"
 import useWarnIfUnsavedChanges from "hooks/useWarnIfUnsavedChanges"
@@ -116,11 +115,13 @@ const EditRole = ({ roleData }: Props): JSX.Element => {
       >
         <DrawerOverlay />
         <DrawerContent>
-          <DrawerCloseButton rounded="full" />
           <DrawerBody className="custom-scrollbar">
-            <Heading as="h3" mb={8} fontFamily="display">
-              Edit role
-            </Heading>
+            <HStack justifyContent="space-between" mb={8}>
+              <Heading as="h3" fontFamily="display">
+                Edit role
+              </Heading>
+              <DeleteRoleButton roleId={id} />
+            </HStack>
             <FormProvider {...methods}>
               <VStack spacing={10} alignItems="start">
                 <Section title="Choose a logo and name for your role">
@@ -136,12 +137,6 @@ const EditRole = ({ roleData }: Props): JSX.Element => {
                 </Section>
 
                 <Requirements maxCols={2} />
-
-                <Divider
-                  borderColor={colorMode === "light" ? "blackAlpha.400" : undefined}
-                />
-
-                <DeleteRoleCard roleId={id} />
               </VStack>
             </FormProvider>
           </DrawerBody>
