+++ src/components/[guild]/edit/[role]/DeleteRoleButton/index.ts
@@ -1,0 +1,3 @@
+import DeleteRoleButton from "./DeleteRoleButton"
+
+export default DeleteRoleButton
