--- src/components/[guild]/edit/[role]/DeleteRoleCard/index.ts
@@ -1,3 +1,0 @@
-import DeleteRoleCard from "./DeleteRoleCard"
-
-export default DeleteRoleCard
