+++ src/components/[guild]/edit/index/DeleteGuildButton/DeleteGuildButton.tsx
@@ -1,0 +1,70 @@
+import {
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  Button,
+  Icon,
+  IconButton,
+  Text,
+  useBreakpointValue,
+  useDisclosure,
+} from "@chakra-ui/react"
+import usePersonalSign from "hooks/usePersonalSign"
+import { TrashSimple } from "phosphor-react"
+import { useRef } from "react"
+import useDeleteGuild from "./hooks/useDeleteGuild"
+
+const DeleteGuildButton = (): JSX.Element => {
+  const { isOpen, onOpen, onClose } = useDisclosure()
+  const { onSubmit, isLoading } = useDeleteGuild()
+  const { isSigning } = usePersonalSign()
+
+  const cancelRef = useRef()
+  const transition = useBreakpointValue<any>({ base: "slideInBottom", sm: "scale" })
+
+  return (
+    <>
+      <IconButton
+        aria-label="Delete guild"
+        icon={<Icon as={TrashSimple} />}
+        colorScheme="red"
+        onClick={onOpen}
+        maxW={10}
+        maxH={10}
+      />
+      <AlertDialog
+        motionPreset={transition}
+        leastDestructiveRef={cancelRef}
+        {...{ isOpen, onClose }}
+      >
+        <AlertDialogOverlay>
+          <AlertDialogContent>
+            <AlertDialogHeader>Delete guild</AlertDialogHeader>
+            <AlertDialogBody>
+              <Text>Are you sure? You can't undo this action afterwards.</Text>
+            </AlertDialogBody>
+            <AlertDialogFooter>
+              <Button ref={cancelRef} onClick={onClose}>
+                Cancel
+              </Button>
+              <Button
+                colorScheme="red"
+                isLoading={isLoading}
+                loadingText={isSigning ? "Check your wallet" : "Deleting"}
+                onClick={() => onSubmit()}
+                ml={3}
+              >
+                Delete
+              </Button>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialogOverlay>
+      </AlertDialog>
+    </>
+  )
+}
+
+export default DeleteGuildButton
