+++ src/components/[guild]/edit/index/DeleteGuildButton/index.ts
@@ -1,0 +1,3 @@
+import DeleteGuildButton from "./DeleteGuildButton"
+
+export default DeleteGuildButton
