--- src/components/[guild]/edit/index/DeleteGuildCard/DeleteGuildCard.tsx
@@ -1,73 +1,0 @@
-import {
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogOverlay,
-  Button,
-  Icon,
-  Text,
-  useBreakpointValue,
-  useDisclosure,
-} from "@chakra-ui/react"
-import Card from "components/common/Card"
-import Section from "components/common/Section"
-import usePersonalSign from "hooks/usePersonalSign"
-import { TrashSimple } from "phosphor-react"
-import { useRef } from "react"
-import useDeleteGuild from "./hooks/useDeleteGuild"
-
-const DeleteGuildCard = (): JSX.Element => {
-  const { isOpen, onOpen, onClose } = useDisclosure()
-  const { onSubmit, isLoading } = useDeleteGuild()
-  const { isSigning } = usePersonalSign()
-
-  const cancelRef = useRef()
-  const transition = useBreakpointValue<any>({ base: "slideInBottom", sm: "scale" })
-
-  return (
-    <Card p="8" w="full">
-      <Section title="Danger zone" alignItems="flex-start">
-        <Button
-          colorScheme="red"
-          variant="outline"
-          onClick={onOpen}
-          leftIcon={<Icon as={TrashSimple} />}
-        >
-          Delete guild
-        </Button>
-        <AlertDialog
-          motionPreset={transition}
-          leastDestructiveRef={cancelRef}
-          {...{ isOpen, onClose }}
-        >
-          <AlertDialogOverlay>
-            <AlertDialogContent>
-              <AlertDialogHeader>Delete guild</AlertDialogHeader>
-              <AlertDialogBody>
-                <Text>Are you sure? You can't undo this action afterwards.</Text>
-              </AlertDialogBody>
-              <AlertDialogFooter>
-                <Button ref={cancelRef} onClick={onClose}>
-                  Cancel
-                </Button>
-                <Button
-                  colorScheme="red"
-                  isLoading={isLoading}
-                  loadingText={isSigning ? "Check your wallet" : "Deleting"}
-                  onClick={() => onSubmit()}
-                  ml={3}
-                >
-                  Delete
-                </Button>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialogOverlay>
-        </AlertDialog>
-      </Section>
-    </Card>
-  )
-}
-
-export default DeleteGuildCard
