--- src/components/[guild]/edit/index/DeleteGuildCard/index.ts
@@ -1,3 +1,0 @@
-import DeleteGuildCard from "./DeleteGuildCard"
-
-export default DeleteGuildCard
